pyproject.toml
[project]
name = "assignment-omni"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Yash", email = "71819535+yashupadhyayy1@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.2.16",
    "langchain-community>=0.2.16",
    "langgraph>=0.2.26",
    "langsmith>=0.1.99",
    "qdrant-client>=1.11.3",
    "pydantic>=2.8.2",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.7.0",
    "pypdf>=4.2.0",
    "sentence-transformers>=3.0.1",
    "streamlit>=1.37.0",
    "fastembed>=0.3.6",
    "openai>=1.40.0",
    "langchain-ollama>=0.1.0",
    "langchain-qdrant>=0.1.0"
]

[project.scripts]
assignment-omni = "assignment_omni:main"

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.1",
    "pytest-asyncio>=0.23.7",
    "ruff>=0.5.5",
    "mypy>=1.10.0",
    "types-requests",
    "pytest-cov>=5.0.0"
]
